requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy[asyncio]
aiosqlite
passlib[bcrypt]
python-jose
pydantic[email]
